Move method call txn creation from ATC to MethodCallParams
