+++ src/main/java/com/algorand/algosdk/transaction/AtomicTransactionComposer.java
@@ -1,14 +1,8 @@
 package com.algorand.algosdk.transaction;
 
 import com.algorand.algosdk.abi.Method;
-import com.algorand.algosdk.abi.TypeAddress;
-import com.algorand.algosdk.abi.TypeUint;
-import com.algorand.algosdk.abi.TypeTuple;
 import com.algorand.algosdk.abi.ABIType;
-import com.algorand.algosdk.builder.transaction.ApplicationCallTransactionBuilder;
-import com.algorand.algosdk.crypto.Address;
 import com.algorand.algosdk.crypto.Digest;
-import com.algorand.algosdk.logic.StateSchema;
 import com.algorand.algosdk.util.Encoder;
 import com.algorand.algosdk.v2.client.Utils;
 import com.algorand.algosdk.v2.client.common.AlgodClient;
@@ -19,7 +13,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.ArrayUtils;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.ArrayList;
@@ -37,10 +30,6 @@ public class AtomicTransactionComposer {
     }
 
     public static final int MAX_GROUP_SIZE = 16;
-    // if the abi type argument number > 15, then the abi types after 14th should be wrapped in a tuple
-    private static final int MAX_ABI_ARG_TYPE_LEN = 15;
-
-    private static final int FOREIGN_OBJ_ABI_UINT_SIZE = 8;
 
     private static final byte[] ABI_RET_HASH = new byte[]{0x15, 0x1f, 0x7c, 0x75};
 
@@ -105,35 +94,6 @@ public class AtomicTransactionComposer {
     }
 
     /**
-     * Add a value to an application call's foreign array. The addition will be as compact as possible,
-     * and this function will return an index that can be used to reference `objectToBeAdded` in `objectArray`.
-     *
-     * @param objectToBeAdded - The value to add to the array. If this value is already present in the array,
-     *   it will not be added again. Instead, the existing index will be returned.
-     * @param objectArray - The existing foreign array. This input may be modified to append `valueToAdd`.
-     * @param zerothObject - If provided, this value indicated two things: the 0 value is special for this
-     *   array, so all indexes into `objectArray` must start at 1; additionally, if `objectToBeAdded` equals
-     *   `zerothValue`, then `objectToBeAdded` will not be added to the array, and instead the 0 indexes will
-     *   be returned.
-     * @return An index that can be used to reference `valueToAdd` in `array`.
-     */
-    private static <T> int populateForeignArrayIndex(T objectToBeAdded, List<T> objectArray, T zerothObject) {
-        if (objectToBeAdded.equals(zerothObject))
-            return 0;
-        int startFrom = zerothObject == null ? 0 : 1;
-        int searchInListIndex = objectArray.indexOf(objectToBeAdded);
-        if (searchInListIndex != -1)
-            return startFrom + searchInListIndex;
-        objectArray.add(objectToBeAdded);
-        return objectArray.size() - 1 + startFrom;
-    }
-
-    private static boolean checkTransactionType(TransactionWithSigner tws, String txnType) {
-        if (txnType.equals(Method.TxAnyType)) return true;
-        return tws.txn.type.toValue().equals(txnType);
-    }
-
-    /**
      * Add a smart contract method call to this atomic group.
      * <p>
      * An error will be thrown if the composer's status is not BUILDING, if adding this transaction
@@ -145,116 +105,15 @@ public class AtomicTransactionComposer {
     public void addMethodCall(MethodCallParams methodCall) {
         if (!this.status.equals(Status.BUILDING))
             throw new IllegalArgumentException("Atomic Transaction Composer must be in BUILDING stage");
-        if (this.transactionList.size() + methodCall.method.getTxnCallCount() > MAX_GROUP_SIZE)
+
+        List<TransactionWithSigner> txns = methodCall.createTransactions();
+
+        if (this.transactionList.size() + txns.size() > MAX_GROUP_SIZE)
             throw new IllegalArgumentException(
                     "Atomic Transaction Composer cannot exceed MAX_GROUP_SIZE = " + MAX_GROUP_SIZE + " transactions"
             );
 
-        List<byte[]> encodedABIArgs = new ArrayList<>();
-        encodedABIArgs.add(methodCall.method.getSelector());
-
-        List<Object> methodArgs = new ArrayList<>();
-        List<ABIType> methodABIts = new ArrayList<>();
-
-        List<TransactionWithSigner> tempTransWithSigner = new ArrayList<>();
-        List<Address> foreignAccounts = new ArrayList<>(methodCall.foreignAccounts);
-        List<Long> foreignAssets = new ArrayList<>(methodCall.foreignAssets);
-        List<Long> foreignApps = new ArrayList<>(methodCall.foreignApps);
-
-        for (int i = 0; i < methodCall.method.args.size(); i++) {
-            Method.Arg argT = methodCall.method.args.get(i);
-            Object methodArg = methodCall.methodArgs.get(i);
-            if (argT.parsedType == null && methodArg instanceof TransactionWithSigner) {
-                TransactionWithSigner twsConverted = (TransactionWithSigner) methodArg;
-                if (!checkTransactionType(twsConverted, argT.type))
-                    throw new IllegalArgumentException(
-                            "expected transaction type " + argT.type
-                                    + " not match with given " + twsConverted.txn.type.toValue()
-                    );
-                tempTransWithSigner.add((TransactionWithSigner) methodArg);
-            } else if (Method.RefArgTypes.contains(argT.type)) {
-                int index;
-                if (argT.type.equals(Method.RefTypeAccount)) {
-                    TypeAddress abiAddressT = new TypeAddress();
-                    Address accountAddress = (Address) abiAddressT.decode(abiAddressT.encode(methodArg));
-                    index = populateForeignArrayIndex(accountAddress, foreignAccounts, methodCall.sender);
-                } else if (argT.type.equals(Method.RefTypeAsset) && methodArg instanceof BigInteger) {
-                    TypeUint abiUintT = new TypeUint(64);
-                    BigInteger assetID = (BigInteger) abiUintT.decode(abiUintT.encode(methodArg));
-                    index = populateForeignArrayIndex(assetID.longValue(), foreignAssets, null);
-                } else if (argT.type.equals(Method.RefTypeApplication) && methodArg instanceof BigInteger) {
-                    TypeUint abiUintT = new TypeUint(64);
-                    BigInteger appID = (BigInteger) abiUintT.decode(abiUintT.encode(methodArg));
-                    index = populateForeignArrayIndex(appID.longValue(), foreignApps, methodCall.appID);
-                } else
-                    throw new IllegalArgumentException(
-                            "cannot add method call in AtomicTransactionComposer: ForeignArray arg type not matching"
-                    );
-                methodArgs.add(index);
-                methodABIts.add(new TypeUint(FOREIGN_OBJ_ABI_UINT_SIZE));
-            } else if (argT.parsedType != null) {
-                methodArgs.add(methodArg);
-                methodABIts.add(argT.parsedType);
-            } else
-                throw new IllegalArgumentException(
-                        "error: the type of method argument is a transaction-type, but no transaction-with-signer provided"
-                );
-        }
-
-        if (methodArgs.size() > MAX_ABI_ARG_TYPE_LEN) {
-            List<ABIType> wrappedABITypeList = new ArrayList<>();
-            List<Object> wrappedValueList = new ArrayList<>();
-
-            for (int i = MAX_ABI_ARG_TYPE_LEN - 1; i < methodArgs.size(); i++) {
-                wrappedABITypeList.add(methodABIts.get(i));
-                wrappedValueList.add(methodArgs.get(i));
-            }
-
-            TypeTuple tupleT = new TypeTuple(wrappedABITypeList);
-            methodABIts = methodABIts.subList(0, MAX_ABI_ARG_TYPE_LEN - 1);
-            methodABIts.add(tupleT);
-            methodArgs = methodArgs.subList(0, MAX_ABI_ARG_TYPE_LEN - 1);
-            methodArgs.add(wrappedValueList);
-        }
-
-        for (int i = 0; i < methodArgs.size(); i++)
-            encodedABIArgs.add(methodABIts.get(i).encode(methodArgs.get(i)));
-
-        ApplicationCallTransactionBuilder<?> txBuilder = ApplicationCallTransactionBuilder.Builder();
-
-        txBuilder
-            .firstValid(methodCall.firstValid)
-            .lastValid(methodCall.lastValid)
-            .genesisHash(methodCall.genesisHash)
-            .genesisID(methodCall.genesisID)
-            .fee(methodCall.fee)
-            .flatFee(methodCall.flatFee)
-            .note(methodCall.note)
-            .lease(methodCall.lease)
-            .rekey(methodCall.rekeyTo)
-            .sender(methodCall.sender)
-            .applicationId(methodCall.appID)
-            .args(encodedABIArgs)
-            .accounts(foreignAccounts)
-            .foreignApps(foreignApps)
-            .foreignAssets(foreignAssets);
-
-        Transaction tx = txBuilder.build();
-
-        // must apply these fields manually, as they are not exposed in the base ApplicationCallTransactionBuilder
-        tx.onCompletion = methodCall.onCompletion;
-        tx.approvalProgram = methodCall.approvalProgram;
-        tx.clearStateProgram = methodCall.clearProgram;
-
-        if (methodCall.globalStateSchema != null)
-            tx.globalStateSchema = methodCall.globalStateSchema;
-        if (methodCall.localStateSchema != null)
-            tx.localStateSchema = methodCall.localStateSchema;
-        if (methodCall.extraPages != null)
-            tx.extraPages = methodCall.extraPages;
-
-        this.transactionList.addAll(tempTransWithSigner);
-        this.transactionList.add(new TransactionWithSigner(tx, methodCall.signer));
+        this.transactionList.addAll(txns);
         this.methodMap.put(this.transactionList.size() - 1, methodCall.method);
     }
