+++ src/main/java/com/algorand/algosdk/transaction/MethodCallParams.java
@@ -1,7 +1,12 @@
 package com.algorand.algosdk.transaction;
 
+import com.algorand.algosdk.abi.ABIType;
 import com.algorand.algosdk.abi.Method;
+import com.algorand.algosdk.abi.TypeAddress;
+import com.algorand.algosdk.abi.TypeTuple;
+import com.algorand.algosdk.abi.TypeUint;
 import com.algorand.algosdk.algod.client.model.TransactionParams;
+import com.algorand.algosdk.builder.transaction.ApplicationCallTransactionBuilder;
 import com.algorand.algosdk.builder.transaction.MethodCallTransactionBuilder;
 import com.algorand.algosdk.crypto.Address;
 import com.algorand.algosdk.crypto.Digest;
@@ -17,6 +22,11 @@ import java.util.List;
  * MethodCallParams is an object that holds all parameters necessary to invoke {@link AtomicTransactionComposer#addMethodCall(MethodCallParams)}
  */
 public class MethodCallParams {
+    // if the abi type argument number > 15, then the abi types after 14th should be wrapped in a tuple
+    private static final int MAX_ABI_ARG_TYPE_LEN = 15;
+
+    private static final int FOREIGN_OBJ_ABI_UINT_SIZE = 8;
+
     public final Long appID;
     public final Transaction.OnCompletion onCompletion;
     public final Method method;
@@ -44,6 +54,10 @@ public class MethodCallParams {
     public final String genesisID;
     public final Digest genesisHash;
 
+    /**
+     * NOTE: it's strongly suggested to use {@link com.algorand.algosdk.builder.transaction.MethodCallTransactionBuilder}
+     * instead of this constructor to create a new MethodCallParams object.
+     */
     public MethodCallParams(Long appID, Method method, List<Object> methodArgs, Address sender,
                             Transaction.OnCompletion onCompletion, byte[] note, byte[] lease, String genesisID, Digest genesisHash,
                             BigInteger firstValid, BigInteger lastValid, BigInteger fee, BigInteger flatFee,
@@ -104,6 +118,150 @@ public class MethodCallParams {
         this.globalStateSchema = globalStateSchema;
         this.localStateSchema = localStateSchema;
         this.extraPages = extraPages;
+    }
+
+    /**
+     * Create the transactions which will carry out the specified method call.
+     * 
+     * The list of transactions returned by this function will have the same length as method.getTxnCallCount().
+     */
+    public List<TransactionWithSigner> createTransactions() {
+        List<byte[]> encodedABIArgs = new ArrayList<>();
+        encodedABIArgs.add(this.method.getSelector());
+
+        List<Object> methodArgs = new ArrayList<>();
+        List<ABIType> methodABIts = new ArrayList<>();
+
+        List<TransactionWithSigner> transactionArgs = new ArrayList<>();
+        List<Address> foreignAccounts = new ArrayList<>(this.foreignAccounts);
+        List<Long> foreignAssets = new ArrayList<>(this.foreignAssets);
+        List<Long> foreignApps = new ArrayList<>(this.foreignApps);
+
+        for (int i = 0; i < this.method.args.size(); i++) {
+            Method.Arg argT = this.method.args.get(i);
+            Object methodArg = this.methodArgs.get(i);
+            if (argT.parsedType == null && methodArg instanceof TransactionWithSigner) {
+                TransactionWithSigner twsConverted = (TransactionWithSigner) methodArg;
+                if (!checkTransactionType(twsConverted, argT.type))
+                    throw new IllegalArgumentException(
+                            "expected transaction type " + argT.type
+                                    + " not match with given " + twsConverted.txn.type.toValue()
+                    );
+                transactionArgs.add((TransactionWithSigner) methodArg);
+            } else if (Method.RefArgTypes.contains(argT.type)) {
+                int index;
+                if (argT.type.equals(Method.RefTypeAccount)) {
+                    TypeAddress abiAddressT = new TypeAddress();
+                    Address accountAddress = (Address) abiAddressT.decode(abiAddressT.encode(methodArg));
+                    index = populateForeignArrayIndex(accountAddress, foreignAccounts, this.sender);
+                } else if (argT.type.equals(Method.RefTypeAsset) && methodArg instanceof BigInteger) {
+                    TypeUint abiUintT = new TypeUint(64);
+                    BigInteger assetID = (BigInteger) abiUintT.decode(abiUintT.encode(methodArg));
+                    index = populateForeignArrayIndex(assetID.longValue(), foreignAssets, null);
+                } else if (argT.type.equals(Method.RefTypeApplication) && methodArg instanceof BigInteger) {
+                    TypeUint abiUintT = new TypeUint(64);
+                    BigInteger appID = (BigInteger) abiUintT.decode(abiUintT.encode(methodArg));
+                    index = populateForeignArrayIndex(appID.longValue(), foreignApps, this.appID);
+                } else
+                    throw new IllegalArgumentException(
+                            "cannot add method call in AtomicTransactionComposer: ForeignArray arg type not matching"
+                    );
+                methodArgs.add(index);
+                methodABIts.add(new TypeUint(FOREIGN_OBJ_ABI_UINT_SIZE));
+            } else if (argT.parsedType != null) {
+                methodArgs.add(methodArg);
+                methodABIts.add(argT.parsedType);
+            } else
+                throw new IllegalArgumentException(
+                        "error: the type of method argument is a transaction-type, but no transaction-with-signer provided"
+                );
+        }
+
+        if (methodArgs.size() > MAX_ABI_ARG_TYPE_LEN) {
+            List<ABIType> wrappedABITypeList = new ArrayList<>();
+            List<Object> wrappedValueList = new ArrayList<>();
+
+            for (int i = MAX_ABI_ARG_TYPE_LEN - 1; i < methodArgs.size(); i++) {
+                wrappedABITypeList.add(methodABIts.get(i));
+                wrappedValueList.add(methodArgs.get(i));
+            }
+
+            TypeTuple tupleT = new TypeTuple(wrappedABITypeList);
+            methodABIts = methodABIts.subList(0, MAX_ABI_ARG_TYPE_LEN - 1);
+            methodABIts.add(tupleT);
+            methodArgs = methodArgs.subList(0, MAX_ABI_ARG_TYPE_LEN - 1);
+            methodArgs.add(wrappedValueList);
+        }
+
+        for (int i = 0; i < methodArgs.size(); i++)
+            encodedABIArgs.add(methodABIts.get(i).encode(methodArgs.get(i)));
+
+        ApplicationCallTransactionBuilder<?> txBuilder = ApplicationCallTransactionBuilder.Builder();
+
+        txBuilder
+            .firstValid(this.firstValid)
+            .lastValid(this.lastValid)
+            .genesisHash(this.genesisHash)
+            .genesisID(this.genesisID)
+            .fee(this.fee)
+            .flatFee(this.flatFee)
+            .note(this.note)
+            .lease(this.lease)
+            .rekey(this.rekeyTo)
+            .sender(this.sender)
+            .applicationId(this.appID)
+            .args(encodedABIArgs)
+            .accounts(foreignAccounts)
+            .foreignApps(foreignApps)
+            .foreignAssets(foreignAssets);
+
+        Transaction tx = txBuilder.build();
+
+        // must apply these fields manually, as they are not exposed in the base ApplicationCallTransactionBuilder
+        tx.onCompletion = this.onCompletion;
+        tx.approvalProgram = this.approvalProgram;
+        tx.clearStateProgram = this.clearProgram;
+
+        if (this.globalStateSchema != null)
+            tx.globalStateSchema = this.globalStateSchema;
+        if (this.localStateSchema != null)
+            tx.localStateSchema = this.localStateSchema;
+        if (this.extraPages != null)
+            tx.extraPages = this.extraPages;
+        
+        TransactionWithSigner methodCall = new TransactionWithSigner(tx, this.signer);
+        transactionArgs.add(methodCall);
+
+        return transactionArgs;
+    }
+
+    private static boolean checkTransactionType(TransactionWithSigner tws, String txnType) {
+        if (txnType.equals(Method.TxAnyType)) return true;
+        return tws.txn.type.toValue().equals(txnType);
+    }
+
+    /**
+     * Add a value to an application call's foreign array. The addition will be as compact as possible,
+     * and this function will return an index that can be used to reference `objectToBeAdded` in `objectArray`.
+     *
+     * @param objectToBeAdded - The value to add to the array. If this value is already present in the array,
+     *   it will not be added again. Instead, the existing index will be returned.
+     * @param objectArray - The existing foreign array. This input may be modified to append `valueToAdd`.
+     * @param zerothObject - If provided, this value indicated two things: the 0 value is special for this
+     *   array, so all indexes into `objectArray` must start at 1; additionally, if `objectToBeAdded` equals
+     *   `zerothValue`, then `objectToBeAdded` will not be added to the array, and instead the 0 indexes will
+     *   be returned.
+     * @return An index that can be used to reference `valueToAdd` in `array`.
+     */
+    private static <T> int populateForeignArrayIndex(T objectToBeAdded, List<T> objectArray, T zerothObject) {
+        if (objectToBeAdded.equals(zerothObject))
+            return 0;
+        int startFrom = zerothObject == null ? 0 : 1;
+        int searchInListIndex = objectArray.indexOf(objectToBeAdded);
+        if (searchInListIndex != -1)
+            return startFrom + searchInListIndex;
+        objectArray.add(objectToBeAdded);
+        return objectArray.size() - 1 + startFrom;
     }
 
     /**
